+++ src/ServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace Kronthto\LaravelOAuth2Login;
 
+use Auth;
 use Illuminate\Support\ServiceProvider as BaseServiceProvider;
 
 class ServiceProvider extends BaseServiceProvider
@@ -17,6 +18,7 @@ class ServiceProvider extends BaseServiceProvider
     {
         $this->publishes([$this->configPath() => config_path('oauth2login.php')]);
         $this->loadRoutesFrom($this->resourcePath().'routes.php');
+        Auth::viaRequest(config('oauth2login.auth_driver_key'), app(AuthFromRequest::class));
     }
 
     /** @return string */