+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- `Auth`-driver *viaRequest* & `Authenticatable` wrapper
 
 ## [1.1.0] - 2017-05-25
 ### Added